backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cloudinary==1.36.0
Django==4.2.6
djangorestframework==3.14.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.7
